grub: move gboot/entry into GRUB
env files are a general grub concept.
